fix: 开源版质量红线相关异常 #1579. fix api path
