+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/api/user/UserPipelineResource.kt
@@ -562,30 +562,6 @@ interface UserPipelineResource {
         pageSize: Int?
     ): Result<PipelineViewPipelinePage<PipelineInfo>>
 
-    @ApiOperation("用户获取视图流水线编排列表-红线")
-    @GET
-    @Path("/projects/{projectId}/listQualityViewPipelines")
-    fun listQualityViewPipelines(
-        @ApiParam(value = "用户ID", required = true, defaultValue = AUTH_HEADER_USER_ID_DEFAULT_VALUE)
-        @HeaderParam(AUTH_HEADER_USER_ID)
-        userId: String,
-        @ApiParam("项目ID", required = true)
-        @PathParam("projectId")
-        projectId: String,
-        @ApiParam("流水线名称关键字", required = false)
-        @QueryParam("keywords")
-        keywords: String?,
-        @ApiParam("第几页", required = false, defaultValue = "1")
-        @QueryParam("page")
-        page: Int?,
-        @ApiParam("每页多少条", required = false, defaultValue = "20")
-        @QueryParam("pageSize")
-        pageSize: Int?,
-        @ApiParam("用户视图ID", required = false)
-        @QueryParam("viewId")
-        viewId: String?
-    ): Result<PipelineViewPipelinePage<QualityPipeline>>
-
     @ApiOperation("校验matrix yaml格式")
     @POST
     @Path("/{projectId}/{pipelineId}/matrix/check")
